requirements.txt
# Python Reddit API Wrapper
praw>=7.8.1

# Data manipulation and analysis
pandas>=2.3.3

# Web scraping
beautifulsoup4>=4.13.5
requests>=2.32.5

# Database
sqlalchemy>=2.0.43

# Environment variables
python-dotenv>=1.1.0
